Add new tasks create_doi_sunet and contribs
Once the initial DOI collection process is complete we know the
population of DOIs we are working with in the dataset. We are also able
to join the data to the authors.csv using either the `orcidid` (for
Dimensions and OpenAlex) or `cap_profile_id` (for sul_pub).

The doi_sunet task will create a mapping of `doi -> [sunetid]` using the
pickle files, sul_pub harvest and the authors.csv.

Once the publications datasets are merged the `doi_sunet` mapping is
used to add the `sunetid` column, and then join with the `authors.csv`.

Closes #33
Closes #34