Account for multiple OpenAlex works with the same DOI
The dois_from_orcid() now returns a list of unique DOIs, instead of an
iterator of possiblly unique DOIs. This allows a failing test to pass.

The test for openalex_publications_from_dois() was relaxed a bit to look
for 231 or more publications, since a lookup for an individual DOI can
sometimes pull back multiple works.

The number of columns for OpenAlex is now 53 because we added
`institution_assertions`.